refactor : hotels data 정리 함수 분리
